refactor(progression): split profile and leaderboard renderers
