+++ bot/features/progression/rendering/__init__.py
@@ -1,1 +1,22 @@
-"""Reusable profile and leaderboard rendering components."""
+"""Progression rendering components."""
+
+from .card_drawer import CardDrawer, ImageRenderer
+from .exceptions import (
+    AvatarBytesMissing,
+    AvatarError,
+    AvatarLoadError,
+)
+from .layouts import LeaderboardLayout, ProfileCardLayout
+from .resources import AssetLoader, FontManager
+
+__all__ = [
+    "AssetLoader",
+    "AvatarBytesMissing",
+    "AvatarError",
+    "AvatarLoadError",
+    "CardDrawer",
+    "FontManager",
+    "ImageRenderer",
+    "LeaderboardLayout",
+    "ProfileCardLayout",
+]
